Port ban and mute listener to discord.py
